requirements.txt
torch>=2.1.0
transformers>=4.37.0
accelerate>=0.27.0
einops>=0.7.0
sentencepiece>=0.1.99
safetensors>=0.4.0
pillow>=10.0.0
matplotlib>=3.7.0 
jupyterlab>=4.0.0